requirements.txt
altgraph==0.17.2
asgiref==3.5.2
auto-py-to-exe==2.23.0
backports.zoneinfo==0.2.1
bootstrap-admin==0.4.4
bottle==0.12.23
bottle-websocket==0.2.9
configparser==5.3.0
Django==3.0.6
django-advanced-filters==2.0.0
django-cors-headers==3.13.0
django-filter==1.1.0
django-rest-framework==0.1.0
django-taggit==1.2.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.2.0
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.3
Pillow==9.2.0
pyinstaller==5.3
pyinstaller-hooks-contrib==2022.10
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.2.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0